Фикс при кнопке Назад может быть файл 0-го размера

+++ src/main/java/ru/gadjini/any2any/service/image/editor/StateFather.java
@@ -12,6 +12,7 @@ import ru.gadjini.any2any.common.MessagesProperties;
 import ru.gadjini.any2any.io.SmartTempFile;
 import ru.gadjini.any2any.job.CommonJobExecutor;
 import ru.gadjini.any2any.model.Any2AnyFile;
+import ru.gadjini.any2any.model.EditMediaContext;
 import ru.gadjini.any2any.model.SendFileContext;
 import ru.gadjini.any2any.model.SendFileResult;
 import ru.gadjini.any2any.service.LocalisationService;
@@ -188,8 +189,13 @@ public class StateFather implements State {
             return;
         }
         try {
-            messageService.sendDocument(new SendFileContext(chatId, new File(state.getCurrentFilePath())));
-            messageService.deleteMessage(chatId, state.getMessageId());
+            File file = new File(state.getCurrentFilePath());
+            if (file.length() > 0) {
+                messageService.sendDocument(new SendFileContext(chatId, new File(state.getCurrentFilePath())));
+                messageService.deleteMessage(chatId, state.getMessageId());
+            } else {
+                messageService.editMessageMedia(new EditMediaContext(chatId, state.getMessageId(), state.getCurrentFileId()));
+            }
             commandStateService.deleteState(chatId, command.getHistoryName());
             LOGGER.debug("Image editor state deleted for user " + chatId);
         } catch (Exception ex) {
@@ -226,8 +232,13 @@ public class StateFather implements State {
 
         if (state != null) {
             try {
-                messageService.sendDocument(new SendFileContext(chatId, new File(state.getCurrentFilePath())));
-                messageService.deleteMessage(chatId, state.getMessageId());
+                File file = new File(state.getCurrentFilePath());
+                if (file.length() > 0) {
+                    messageService.sendDocument(new SendFileContext(chatId, new File(state.getCurrentFilePath())));
+                    messageService.deleteMessage(chatId, state.getMessageId());
+                } else {
+                    messageService.editMessageMedia(new EditMediaContext(chatId, state.getMessageId(), state.getCurrentFileId()));
+                }
             } catch (Exception ex) {
                 LOGGER.error(ex.getMessage(), ex);
             }
